Move luasnip config to seperate file
